filter cart routes with the user id

+++ src/controllers/cartController.js
@@ -21,8 +21,13 @@ async function addCartProducts(req, res) {
 }
 
 async function getCartProducts(req, res) {
+  const { user_id } = req.params;
+
   try {
-    const products = await db.collection("cart").find().toArray();
+    const products = await db
+      .collection("cart")
+      .find({ _id: new ObjectId(user_id) })
+      .toArray();
     res.status(200).send(products);
   } catch (error) {
     return res.status(500).send(error.message);
@@ -41,8 +46,10 @@ async function removeCartProducts(req, res) {
 }
 
 async function cleanCart(req, res) {
+  const { user_id } = req.params;
+
   try {
-    await db.collection("cart").deleteMany({});
+    await db.collection("cart").deleteMany({ _id: new ObjectId(user_id) });
     res.sendStatus(200);
   } catch (error) {
     return res.status(500).send(error.message);
